+++ kontrol-frontend/src/components/CytoscapeGraph/plugins/tippy.ts
@@ -1,8 +1,9 @@
 import cytoscapePopper from "cytoscape-popper";
 import tippy, { Instance } from "tippy.js";
+import { NodeVersion } from "@/types";
 import "./tippy.css";
 
-// @ts-expect-error WIP
+// @ts-expect-error Poor typing upstream
 const tippyFactory: cytoscapePopper.PopperFactory = (ref, content) => {
   // Since tippy constructor requires DOM element/elements, create a placeholder
   const dummyDomEle = document.createElement("div");
@@ -14,7 +15,7 @@ const tippyFactory: cytoscapePopper.PopperFactory = (ref, content) => {
     content: content,
     // your own preferences:
     arrow: true,
-    placement: "right",
+    placement: "bottom",
     hideOnClick: false,
     sticky: "reference",
 
@@ -28,24 +29,23 @@ const tippyFactory: cytoscapePopper.PopperFactory = (ref, content) => {
 
 export const createTooltip = (
   node: cytoscape.NodeSingular,
-): Instance | null => {
-  const versions = node.data("versions");
+): { instance: Instance; element: HTMLElement } | null => {
+  const versions: NodeVersion[] = node.data("versions");
   if (!versions || versions.length === 0) {
     return null;
   }
-  const tip = node.popper({
-    content: () => {
-      const elem = document.createElement("div");
-      // TODO: sanitize
-      elem.innerHTML = `Versions: <ul>${versions.map((v: string) => `<li>${v.toString()}</li>`).join("")}</ul>`;
-      elem.classList.add("tooltip");
-      return elem;
-    },
+
+  const element = document.createElement("div");
+  element.classList.add("tooltip");
+
+  const instance = node.popper({
+    // tooltip content is rendered through the TooltipPortal component
+    content: () => element,
   }) as unknown as Instance;
 
-  tip.show();
+  instance.show();
 
-  return tip;
+  return { instance, element };
 };
 
 export type TooltipInstance = Instance;