+++ kontrol-frontend/src/components/Navbar.tsx
@@ -7,9 +7,21 @@ import {
   Image,
   Spacer,
   AvatarBadge,
+  Menu,
+  MenuButton,
+  MenuList,
+  MenuItem,
+  Button,
 } from "@chakra-ui/react";
+import { useNavigate } from "react-router-dom";
 
 const Navbar = () => {
+  const navigate = useNavigate();
+
+  const navigateToPlaceholder = () => {
+    navigate("profile");
+  };
+
   return (
     <Box
       bg="white"
@@ -21,7 +33,6 @@ const Navbar = () => {
       borderBottomColor="gray.100"
     >
       <Flex h={16} alignItems="center">
-        {/* Logo */}
         <Link href="/" display="flex" alignItems="center">
           <Image
             src="/logo.png"
@@ -36,17 +47,19 @@ const Navbar = () => {
 
         <Spacer />
 
-        {/* User Profile */}
         <Flex alignItems="center">
-          <Avatar size="sm" name="Charlie Brown">
-            <AvatarBadge boxSize="1.25em" bg="green.500" />
-          </Avatar>
-          <Box ml={2}>
-            <Text>Charlie Brown</Text>
-            <Text fontSize="sm" color="gray.500">
-              ACME Corporation
-            </Text>
-          </Box>
+          <Menu>
+            <MenuButton as={Button} p={0} bg={"white"}>
+              <Avatar size="sm">
+                <AvatarBadge boxSize="1.25em" bg="green.500" />
+              </Avatar>
+            </MenuButton>
+            <MenuList>
+              <MenuItem onClick={navigateToPlaceholder}>Profile</MenuItem>
+              <MenuItem onClick={navigateToPlaceholder}>Settings</MenuItem>
+              <MenuItem onClick={navigateToPlaceholder}>Log in</MenuItem>
+            </MenuList>
+          </Menu>
         </Flex>
       </Flex>
     </Box>