+++ kontrol-service/constants/constants.go
@@ -2,4 +2,5 @@ package constants
 
 const (
 	SharedVersionVersionString = "shared"
+	DefaultNS                  = "default"
 )