+++ kontrol-service/engine/flow/dev_flow.go
@@ -2,6 +2,7 @@ package flow
 
 import (
 	"fmt"
+
 	"kardinal.kontrol-service/constants"
 
 	"github.com/kurtosis-tech/stacktrace"
@@ -20,7 +21,12 @@ import (
 // CreateDevFlow creates a dev flow from the given topologies
 // baseClusterTopologyMaybeWithTemplateOverrides - if a template is used then this is a modified version of the baseTopology
 // we pass in the base topology anyway as we use services which remain in `prod` version from it
-func CreateDevFlow(pluginRunner *plugins.PluginRunner, baseClusterTopologyMaybeWithTemplateOverrides resolved.ClusterTopology, baseTopology resolved.ClusterTopology, flowPatch flow_spec.FlowPatch) (*resolved.ClusterTopology, error) {
+func CreateDevFlow(
+	pluginRunner *plugins.PluginRunner,
+	baseClusterTopologyMaybeWithTemplateOverrides resolved.ClusterTopology,
+	baseTopology resolved.ClusterTopology,
+	flowPatch flow_spec.FlowPatch,
+) (*resolved.ClusterTopology, error) {
 	flowID := flowPatch.FlowId
 
 	// shallow copy the base topology
@@ -30,15 +36,15 @@ func CreateDevFlow(pluginRunner *plugins.PluginRunner, baseClusterTopologyMaybeW
 	topology.FlowID = flowID
 	topology.Services = deepCopySlice(baseClusterTopologyMaybeWithTemplateOverrides.Services)
 	topology.ServiceDependencies = deepCopySlice(baseClusterTopologyMaybeWithTemplateOverrides.ServiceDependencies)
-	topology.Ingresses = lo.Map(baseClusterTopologyMaybeWithTemplateOverrides.Ingresses, func(item *resolved.Ingress, _ int) *resolved.Ingress {
-		copiedIngress := resolved.Ingress{
-			ActiveFlowIDs: []string{flowID},
-			IngressID:     item.IngressID,
-			IngressRules:  deepCopySlice(item.IngressRules),
-			ServiceSpec:   item.ServiceSpec,
-		}
-		return &copiedIngress
-	})
+	topology.Ingress = &resolved.Ingress{
+		ActiveFlowIDs: []string{flowID},
+		Ingresses:     deepCopySlice(baseClusterTopologyMaybeWithTemplateOverrides.Ingress.Ingresses),
+	}
+	topology.GatewayAndRoutes = &resolved.GatewayAndRoutes{
+		ActiveFlowIDs: []string{flowID},
+		Gateways:      deepCopySlice(baseClusterTopologyMaybeWithTemplateOverrides.GatewayAndRoutes.Gateways),
+		GatewayRoutes: deepCopySlice(baseClusterTopologyMaybeWithTemplateOverrides.GatewayAndRoutes.GatewayRoutes),
+	}
 
 	topologyRef := &topology
 
@@ -138,7 +144,7 @@ func markParentsAsShared(topology *resolved.ClusterTopology, service *resolved.S
 func applyPatch(
 	pluginRunner *plugins.PluginRunner,
 	topology *resolved.ClusterTopology,
-	clusterGraph graph.Graph[*resolved.Service, *resolved.Service],
+	clusterGraph graph.Graph[resolved.ServiceHash, *resolved.Service],
 	flowID string,
 	targetService *resolved.Service,
 	deploymentSpec *v1.DeploymentSpec,
@@ -147,9 +153,13 @@ func applyPatch(
 	statefulPaths := findAllDownstreamStatefulPaths(targetService, clusterGraph, topology)
 	statefulServices := make([]*resolved.Service, 0)
 	for _, path := range statefulPaths {
-		statefulService, err := lo.Last(path)
-		if statefulService == nil || err != nil {
-			logrus.Infof("Error finding last stateful service in path %v: %v", path, err)
+		statefulServiceHash, err := lo.Last(path)
+		if statefulServiceHash == "" || err != nil {
+			logrus.Infof("Error finding last stateful service hash in path %v: %v", path, err)
+		}
+		statefulService, err := clusterGraph.Vertex(statefulServiceHash)
+		if err != nil {
+			return nil, fmt.Errorf("an error occurred getting stateful service vertex from graph: %s", err)
 		}
 		statefulServices = append(statefulServices, statefulService)
 	}
@@ -159,9 +169,13 @@ func applyPatch(
 	externalServices := make([]*resolved.Service, 0)
 	alreadyHandledExternalServices := make([]string, 0)
 	for _, path := range externalPaths {
-		externalService, err := lo.Last(path)
-		if externalService == nil || err != nil {
-			logrus.Infof("Error finding last external service in path %v: %v", path, err)
+		externalServiceHash, err := lo.Last(path)
+		if externalServiceHash == "" || err != nil {
+			logrus.Infof("Error finding last external service hash in path %v: %v", path, err)
+		}
+		externalService, err := clusterGraph.Vertex(externalServiceHash)
+		if err != nil {
+			return nil, fmt.Errorf("an error occurred getting external service vertex from graph: %s", err)
 		}
 		externalServices = append(externalServices, externalService)
 	}
@@ -302,7 +316,8 @@ func applyExternalServicePlugin(
 	externalService *resolved.Service,
 	externalServicePlugin *resolved.StatefulPlugin,
 	pluginIdx int,
-	flowId string) error {
+	flowId string,
+) error {
 	if externalServicePlugin.Type != "external" {
 		return nil
 	}
@@ -345,9 +360,9 @@ func DeleteDevFlow(pluginRunner *plugins.PluginRunner, flowId string, service *r
 	return nil
 }
 
-func topologyToGraph(topology *resolved.ClusterTopology) graph.Graph[*resolved.Service, *resolved.Service] {
-	serviceHash := func(service *resolved.Service) *resolved.Service {
-		return service
+func topologyToGraph(topology *resolved.ClusterTopology) graph.Graph[resolved.ServiceHash, *resolved.Service] {
+	serviceHash := func(service *resolved.Service) resolved.ServiceHash {
+		return service.Hash()
 	}
 	graph := graph.New(serviceHash, graph.Directed())
 
@@ -356,20 +371,20 @@ func topologyToGraph(topology *resolved.ClusterTopology) graph.Graph[*resolved.S
 	}
 
 	for _, dependency := range topology.ServiceDependencies {
-		graph.AddEdge(dependency.Service, dependency.DependsOnService)
+		graph.AddEdge(dependency.Service.Hash(), dependency.DependsOnService.Hash())
 	}
 
 	return graph
 }
 
-func findAllDownstreamStatefulPaths(targetService *resolved.Service, clusterGraph graph.Graph[*resolved.Service, *resolved.Service], topology *resolved.ClusterTopology) [][]*resolved.Service {
-	allPaths := make([][]*resolved.Service, 0)
+func findAllDownstreamStatefulPaths(targetService *resolved.Service, clusterGraph graph.Graph[resolved.ServiceHash, *resolved.Service], topology *resolved.ClusterTopology) [][]resolved.ServiceHash {
+	allPaths := make([][]resolved.ServiceHash, 0)
 	for _, service := range topology.Services {
 		if service.IsStateful {
-			paths, err := graph.AllPathsBetween(clusterGraph, targetService, service)
+			paths, err := graph.AllPathsBetween(clusterGraph, targetService.Hash(), service.Hash())
 			if err != nil {
 				logrus.Infof("Error finding paths between %s and %s: %v", targetService.ServiceID, service.ServiceID, err)
-				paths = [][]*resolved.Service{}
+				paths = [][]resolved.ServiceHash{}
 			}
 			allPaths = append(allPaths, paths...)
 		}
@@ -377,14 +392,14 @@ func findAllDownstreamStatefulPaths(targetService *resolved.Service, clusterGrap
 	return allPaths
 }
 
-func findAllDownstreamExternalPaths(targetService *resolved.Service, g graph.Graph[*resolved.Service, *resolved.Service], topology *resolved.ClusterTopology) [][]*resolved.Service {
-	allPaths := make([][]*resolved.Service, 0)
+func findAllDownstreamExternalPaths(targetService *resolved.Service, g graph.Graph[resolved.ServiceHash, *resolved.Service], topology *resolved.ClusterTopology) [][]resolved.ServiceHash {
+	allPaths := make([][]resolved.ServiceHash, 0)
 	for _, service := range topology.Services {
 		if service.IsExternal {
-			paths, err := graph.AllPathsBetween(g, targetService, service)
+			paths, err := graph.AllPathsBetween(g, targetService.Hash(), service.Hash())
 			if err != nil {
 				logrus.Infof("Error finding paths between %s and %s: %v", targetService.ServiceID, service.ServiceID, err)
-				paths = [][]*resolved.Service{}
+				paths = [][]resolved.ServiceHash{}
 			}
 			allPaths = append(allPaths, paths...)
 		}